+++ main.go
@@ -10,9 +10,9 @@ import (
 	"runtime/debug"
 
 	"github.com/fatih/color"
+	"github.com/git-town/git-town/v9/src/cli/dialog"
 	"github.com/git-town/git-town/v9/src/cli/print"
 	"github.com/git-town/git-town/v9/src/cmd"
-	"github.com/git-town/git-town/v9/src/dialog"
 )
 
 func main() {
