+++ src/cli/dialog/core.go
