+++ src/cli/dialog/select_squash_commit_author.go
