+++ src/cli/flags/debug_test.go
@@ -3,20 +3,20 @@ package flags_test
 import (
 	"testing"
 
-	"github.com/git-town/git-town/v9/src/flags"
+	"github.com/git-town/git-town/v9/src/cli/flags"
 	"github.com/shoenig/test/must"
 	"github.com/spf13/cobra"
 )
 
-func TestBool(t *testing.T) {
+func TestDebug(t *testing.T) {
 	t.Parallel()
 
 	t.Run("long version", func(t *testing.T) {
 		t.Parallel()
 		cmd := cobra.Command{}
-		addFlag, readFlag := flags.Bool("myflag", "m", "desc")
+		addFlag, readFlag := flags.Debug()
 		addFlag(&cmd)
-		err := cmd.ParseFlags([]string{"--myflag"})
+		err := cmd.ParseFlags([]string{"--debug"})
 		must.NoError(t, err)
 		must.EqOp(t, true, readFlag(&cmd))
 	})
@@ -24,9 +24,9 @@ func TestBool(t *testing.T) {
 	t.Run("short version", func(t *testing.T) {
 		t.Parallel()
 		cmd := cobra.Command{}
-		addFlag, readFlag := flags.Bool("myflag", "m", "desc")
+		addFlag, readFlag := flags.Debug()
 		addFlag(&cmd)
-		err := cmd.ParseFlags([]string{"-m"})
+		err := cmd.ParseFlags([]string{"-d"})
 		must.NoError(t, err)
 		must.EqOp(t, true, readFlag(&cmd))
 	})
