+++ src/cmd/config_perennial_branches.go
@@ -1,11 +1,11 @@
 package cmd
 
 import (
+	"github.com/git-town/git-town/v9/src/cli/dialog"
+	"github.com/git-town/git-town/v9/src/cli/flags"
 	"github.com/git-town/git-town/v9/src/cli/format"
 	"github.com/git-town/git-town/v9/src/cli/io"
-	"github.com/git-town/git-town/v9/src/dialog"
 	"github.com/git-town/git-town/v9/src/execute"
-	"github.com/git-town/git-town/v9/src/flags"
 	"github.com/spf13/cobra"
 )
