+++ src/cmd/config_reset.go
@@ -1,8 +1,8 @@
 package cmd
 
 import (
+	"github.com/git-town/git-town/v9/src/cli/flags"
 	"github.com/git-town/git-town/v9/src/execute"
-	"github.com/git-town/git-town/v9/src/flags"
 	"github.com/spf13/cobra"
 )
