+++ src/cmd/status_reset.go
@@ -3,8 +3,8 @@ package cmd
 import (
 	"fmt"
 
+	"github.com/git-town/git-town/v9/src/cli/flags"
 	"github.com/git-town/git-town/v9/src/execute"
-	"github.com/git-town/git-town/v9/src/flags"
 	"github.com/git-town/git-town/v9/src/vm/persistence"
 	"github.com/spf13/cobra"
 )
