+++ src/cmd/switch.go
@@ -3,11 +3,11 @@ package cmd
 import (
 	"strings"
 
+	"github.com/git-town/git-town/v9/src/cli/dialog"
+	"github.com/git-town/git-town/v9/src/cli/flags"
 	"github.com/git-town/git-town/v9/src/config"
-	"github.com/git-town/git-town/v9/src/dialog"
 	"github.com/git-town/git-town/v9/src/domain"
 	"github.com/git-town/git-town/v9/src/execute"
-	"github.com/git-town/git-town/v9/src/flags"
 	"github.com/spf13/cobra"
 )
