+++ src/validate/handle_unfinished_state.go
@@ -3,8 +3,8 @@ package validate
 import (
 	"fmt"
 
+	"github.com/git-town/git-town/v9/src/cli/dialog"
 	"github.com/git-town/git-town/v9/src/config"
-	"github.com/git-town/git-town/v9/src/dialog"
 	"github.com/git-town/git-town/v9/src/domain"
 	"github.com/git-town/git-town/v9/src/git"
 	"github.com/git-town/git-town/v9/src/hosting"
