+++ src/validate/is_configured.go
@@ -3,7 +3,7 @@ package validate
 import (
 	"fmt"
 
-	"github.com/git-town/git-town/v9/src/dialog"
+	"github.com/git-town/git-town/v9/src/cli/dialog"
 	"github.com/git-town/git-town/v9/src/domain"
 	"github.com/git-town/git-town/v9/src/git"
 )
