+++ src/validate/knows_branch_ancestry.go
@@ -1,8 +1,8 @@
 package validate
 
 import (
+	"github.com/git-town/git-town/v9/src/cli/dialog"
 	"github.com/git-town/git-town/v9/src/cli/io"
-	"github.com/git-town/git-town/v9/src/dialog"
 	"github.com/git-town/git-town/v9/src/domain"
 	"github.com/git-town/git-town/v9/src/git"
 )
