+++ src/vm/opcode/squash_merge.go
@@ -3,7 +3,7 @@ package opcode
 import (
 	"fmt"
 
-	"github.com/git-town/git-town/v9/src/dialog"
+	"github.com/git-town/git-town/v9/src/cli/dialog"
 	"github.com/git-town/git-town/v9/src/domain"
 	"github.com/git-town/git-town/v9/src/messages"
 	"github.com/git-town/git-town/v9/src/vm/shared"
